feat: Refactor Users api view This PR refactors the users api method in mysql backend to have all the db logic placed in models instead of the api
