+++ forum/views/comments.py
@@ -24,7 +24,7 @@ class CommentsAPIView(APIView):
 
     permission_classes = (AllowAny,)
 
-    def get(self, request: Request, comment_id: str) -> Response:
+    def get(self, request: Request, comment_id: str) -> Response:  # pylint: disable=W0613
         """
         Retrieves a parent comment.
         For chile comments, below API is called that return all child comments in children field
@@ -133,7 +133,7 @@ class CommentsAPIView(APIView):
             )
         return Response(comment, status=status.HTTP_200_OK)
 
-    def delete(self, request: Request, comment_id: str) -> Response:
+    def delete(self, request: Request, comment_id: str) -> Response:  # pylint: disable=W0613
         """
         Deletes a comment.
